Added Hugo Monteiro's external user lookup patch
Allows dspam to lookup a username for the dspam
database in an external source. This extends the existing
ldap_client interface, and also provides a generic interface for scripted
programs. More detail and the latest versions. This allows the administrator
to define custom usernames to be used, not only system usernames or email addresses.