Don't exit() if malloc fails.
When a library function calls exit(), it prevents the calling program from
handling the error, reporting it to the user, closing files properly, and
cleaning up any state that the program has. It is preferred for the
library to return an actual error code and let the calling program decide
how to handle the situation.

MFM:		21 day
Bug-ID:		2796340
Submitted by:	Stevan Bajic for Andreas Schneider
Security:	none