SQLite3 driver, purge script and documentation update
Change the SQLite3 driver to be closer to the changes recently done in the MySQL and PostgreSQL driver. The changes for the SQLite3 driver include:
* Set all statistical counters to be unsigned
* Fix memory leaks
* Reduce memory footprint
* Fix issues with escaped SQL queries
* Add more debug output (only if debug is enabled)
* Micro speed and memory consumption improvements by removing not needed whitespace from SQL queries.
* Update SQLite3 purge script
* Update documentation for SQLite driver

Database schema has not changed. Users from pre 3.9.0 can just use the new driver without the need to run migration commands on the database.

MFM: 21 days
Bug-ID: 2802205
Bug-Reported-By: M.B. <leto_i@users.sourceforge.net>
Security: none