Send always a final '.' in DSPAM LMTP mode
The 'dspamc' clients expects in summary and classify mode that server
finishes its response with a '.'.  In old logic, this dot was sent
only when there was a response. But when e.g. user could not be found
by external lookup no response and not final dot was sent so that
client hangs until a timeout.

This patches sets the corresponding flag earlier so that '.' is always
sent.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>