don't read past the end of css files
Currently, dspam's code assumes css files are well formed. It will thus
check wether a record exist or not by comparing its position against
the maximum number of records stored in the header. In the case of a
corrupted header, that check could return true and the code access data
beyond the end of the file, leading to a segmentation fault. This patch
enhance cssclean to fix this kind of corruption.

Origin: vendor
Bug: https://sourceforge.net/p/dspam/bug-tracker/170/
Bug-Debian: http://bugs.debian.org/722485
Last-Update: 2013-11-26